Add files via upload
This code is a reference code to import the trained reinforcement learning network, pass inputs from the real robot through the network, and pass the predicted motor commands from the model to the real robot. The input in this case is image from a camera and the network used is a CNN. For our case, we need lidar scan data for input and the model is DQN-based.